+++ build_modules/test/fixtures/a/lib/a_imports_b_no_cycle.dart
@@ -1,0 +1,2 @@
+// ignore_for_file: unused_import
+import 'package:b/b_imports_a_no_cycle.dart';